feat: Added validation to phone number and email fields
